Fix implicit size_t -> int conversions
MSVC2015 warns about these when building for x64.

+++ src/hunspell/affentry.hxx
@@ -118,7 +118,7 @@ class PfxEntry : public AffEntry {
   const char* getKey() { return appnd.c_str(); }
   std::string add(const char* word, size_t len);
 
-  inline short getKeyLen() { return appnd.size(); }
+  inline short getKeyLen() { return short(appnd.size()); }
 
   inline const char* getMorph() { return morphcode; }
 
@@ -199,7 +199,7 @@ class SfxEntry : public AffEntry {
   inline short getContLen() { return contclasslen; }
   inline const char* getAffix() { return appnd.c_str(); }
 
-  inline short getKeyLen() { return appnd.size(); }
+  inline short getKeyLen() { return short(appnd.size()); }
 
   inline SfxEntry* getNext() { return next; }
   inline SfxEntry* getNextNE() { return nextne; }
